Fixed project file for compiling on Windows with llvm 3.4.
- Fixed CastCostCalculator.
- Fixed Function parameter saving to variables.
- Fixed tokenization of string literals.
- Added max(float, int) and max(int, float) (same for min) runtime functions.
- Fixed greaterEqual operator generation.

+++ Compiler/abstractsyntaxtree.h
@@ -756,7 +756,7 @@ class SingleConditionBlockStatement : public BlockNode {
 		static Type staticType() { return NT; }
 		Type type() const { return staticType(); }
 		int childNodeCount() const { return 2; }
-		Node *childNode(int n) const {assert((n == 0 || n == 1) && "Invalid child node id"); return (n == CondBeforeBlock) ? mBlock : mCondition;}
+		Node *childNode(int n) const {assert((n == 0 || n == 1) && "Invalid child node id"); return (n == (int)CondBeforeBlock) ? mBlock : mCondition;}
 
 		Node *condition() const { return mCondition; }
 		void setCondition(Node *n) { mCondition = n; }
